+++ packages/types-core/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add generic network module types
+
 ### Changed
 - Reader interface not returning undefined