+++ packages/types-core/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Add generic network module types
+- Add generic network module types (#2462)
 
 ### Changed
 - Reader interface not returning undefined